create shared template function for tracker networks text
